Move toGuard to guards.ts
