+++ packages/core/src/utils.ts
@@ -323,39 +323,6 @@ export function isString(value: any): value is string {
   return typeof value === 'string';
 }
 
-export function toGuard<TContext, TEvent extends EventObject>(
-  guardConfig: GuardConfig<TContext, TEvent>,
-  guardMap?: Record<string, GuardPredicate<TContext, TEvent>>
-): GuardDefinition<TContext, TEvent> {
-  if (isString(guardConfig)) {
-    return {
-      type: guardConfig,
-      predicate: guardMap ? guardMap[guardConfig] : undefined,
-      params: { type: guardConfig }
-    };
-  }
-
-  if (isFunction(guardConfig)) {
-    return {
-      type: guardConfig.name,
-      predicate: guardConfig,
-      params: {
-        type: guardConfig.name,
-        name: guardConfig.name
-      }
-    };
-  }
-
-  return {
-    type: guardConfig.type,
-    params: guardConfig.params || guardConfig,
-    children: guardConfig.children?.map((childGuard) =>
-      toGuard(childGuard, guardMap)
-    ),
-    predicate: guardConfig.predicate || guardMap?.[guardConfig.type]
-  };
-}
-
 export function isObservable<T>(value: any): value is Subscribable<T> {
   try {
     return 'subscribe' in value && isFunction(value.subscribe);
